Check for errors when printing geom.

+++ include/mid.h
@@ -551,6 +551,6 @@ _Bool scangeom(char *buf, char *fmt, ...);
 /* Prints a structure to a string buffer using the same type of format
  * specified as is used by scanbuf.  The return value is true if the
  * output was not truncated and false if the output was truncated. */
-_Bool printgeom(char *buf, size_t sz, char *fmt, ...);
+_Bool printgeom(char *buf, int sz, char *fmt, ...);
 
 _Bool fsexists(const char *path);
